revert restructuring
for the time being move everything form trixi_version.c to trixi.c
