+++ src/trixi.c
@@ -25,12 +25,32 @@ static void update_depot_path(const char * project_directory, const char * depot
 // `LIBTRIXI_JULIA_DEPOT` in `utils/libtrixi-init-julia` accordingly
 static const char* default_depot_path = "julia-depot";
 
-// List of function pointers, stored below
-void* trixi_function_pointers[TRIXI_NUM_FPTRS];
+// Store function pointers to avoid overhead of `jl_eval_string`
+enum {
+  TRIXI_FTPR_INITIALIZE_SIMULATION,
+  TRIXI_FTPR_CALCULATE_DT,
+  TRIXI_FTPR_IS_FINISHED,
+  TRIXI_FTPR_STEP,
+  TRIXI_FTPR_FINALIZE_SIMULATION,
+  TRIXI_FTPR_NDIMS,
+  TRIXI_FTPR_NELEMENTS,
+  TRIXI_FTPR_NVARIABLES,
+  TRIXI_FTPR_LOAD_CELL_AVERAGES,
+  TRIXI_FTPR_VERSION,
+  TRIXI_FTPR_VERSION_MAJOR,
+  TRIXI_FTPR_VERSION_MINOR,
+  TRIXI_FTPR_VERSION_PATCH,
+  TRIXI_FTPR_VERSION_JULIA,
+  TRIXI_FTPR_VERSION_JULIA_EXTENDED,
+
+  // The last one is for the array size
+  TRIXI_NUM_FPTRS
+};
+static void* trixi_function_pointers[TRIXI_NUM_FPTRS];
 
 // List of function names to obtain C function pointer from Julia
 // OBS! If any name is longer than 250 characters, adjust buffer sizes below
-const char* trixi_function_pointer_names[] = {
+static const char* trixi_function_pointer_names[] = {
   [TRIXI_FTPR_INITIALIZE_SIMULATION]  = "trixi_initialize_simulation_cfptr",
   [TRIXI_FTPR_CALCULATE_DT]           = "trixi_calculate_dt_cfptr",
   [TRIXI_FTPR_IS_FINISHED]            = "trixi_is_finished_cfptr",
@@ -314,6 +334,90 @@ void trixi_load_cell_averages(double * data, int handle) {
 
     // Call function
     load_cell_averages(data, handle);
+}
+
+
+/**
+ * @anchor trixi_version_major_api_c
+ *
+ * @brief Return major version number of libtrixi.
+ *
+ * This function may be run before `trixi_initialize` has been called.
+ *
+ * @return Major version of libtrixi.
+ */
+int trixi_version_major() {
+
+    // Get function pointer
+    int (*version_major)() = trixi_function_pointers[TRIXI_FTPR_VERSION_MAJOR];
+
+    // Call function
+    return version_major();
+}
+
+
+/**
+ * @anchor trixi_version_minor_api_c
+ *
+ * @brief Return minor version number of libtrixi.
+ *
+ * This function may be run before `trixi_initialize` has been called.
+ *
+ * @return Minor version of libtrixi.
+ */
+int trixi_version_minor() {
+
+    // Get function pointer
+    int (*version_minor)() = trixi_function_pointers[TRIXI_FTPR_VERSION_MINOR];
+
+    // Call function
+    return version_minor();
+}
+
+
+/**
+ * @anchor trixi_version_patch_api_c
+ *
+ * @brief Return patch version number of libtrixi.
+ *
+ * This function may be run before `trixi_initialize` has been called.
+ *
+ * @return Patch version of libtrixi.
+ */
+int trixi_version_patch() {
+
+    // Get function pointer
+    int (*version_patch)() = trixi_function_pointers[TRIXI_FTPR_VERSION_PATCH];
+
+    // Call function
+    return version_patch();
+}
+
+
+/**
+ * @anchor trixi_version_api_c
+ *
+ * @brief Return full version string of libtrixi.
+ *
+ * The return value is a read-only pointer to a NULL-terminated string with the version
+ * information. This may include not just MAJOR.MINOR.PATCH but possibly also additional
+ * build or development version information.
+ *
+ * The returned pointer is to static memory and must not be used to change the contents of
+ * the version string. Multiple calls to the function will return the same address.
+ *
+ * This function is thread-safe and may be run before `trixi_initialize` has been called.
+ *
+ * @return Pointer to a read-only, NULL-terminated character array with the full version of
+ *         libtrixi.
+ */
+const char* trixi_version() {
+
+    // Get function pointer
+    const char* (*version)() = trixi_function_pointers[TRIXI_FTPR_VERSION];
+
+    // Call function
+    return version();
 }
